requirements.txt
numpy
pandas
scikit-learn
tensorflow
opencv-python
matplotlib==3.8.0
requests
mediapipe
plotly
fastapi
uvicorn
flask
firebase-admin
google-cloud-firestore
torch
torchvision
Pillow
python-dotenv
pydantic
